requirements.txt
python-dotenv
tensorflow
Flask==1.0.2
grpcio
protobuf